+++ torch/ao/quantization/backend_config/tensorrt.py
@@ -4,7 +4,7 @@ import torch.nn.qat as nnqat
 import torch.nn.intrinsic as nni
 import torch.nn.intrinsic.qat as nniqat
 
-from ...fuser_method_mappings import reverse_sequential_wrapper2
+from ..fuser_method_mappings import reverse_sequential_wrapper2
 
 def get_tensorrt_backend_config_dict():
     """ Get the backend config dictionary for tensorrt backend
