+++ torch/ao/quantization/fx/fuse_handler.py
@@ -1,4 +1,4 @@
-from ..fusion_patterns import DefaultFuseHandler
+from .fusion_patterns import DefaultFuseHandler
 
 # TODO: move DefaultFuseHandler
 def get_fuse_handler_cls():
